[:feat] Add default autodocs setup to preview.tsx
