+++ apps/gallery/.storybook/preview.tsx
@@ -1,0 +1,32 @@
+import { Controls, Description, Primary, Source, Stories, Subtitle, Title } from '@storybook/blocks'
+import React from 'react'
+
+import {} from '@storybook/blocks'
+
+/** @type { import('@storybook/web-components').Preview } */
+const preview = {
+  parameters: {
+    actions: { argTypesRegex: '^on[A-Z].*' },
+    controls: {
+      matchers: {
+        color: /(background|color)$/i,
+        date: /Date$/
+      }
+    },
+    docs: {
+      page: () => (
+        <>
+          <Title />
+          <Subtitle />
+          <Description />
+          <Primary />
+          <Source />
+          <Controls />
+          <Stories />
+        </>
+      )
+    }
+  }
+}
+
+export default preview
