--- apps/gallery/stories/Introduction.mdx
@@ -1,1 +1,0 @@
-Hello Gallery
