+++ apps/gallery/stories/components/wui-text.stories.ts
@@ -10,15 +10,10 @@ export default {
   parameters: {
     layout: 'centered'
   },
-  tags: ['autodocs'],
-
-  render: args =>
-    html`<wui-text variant=${args.variant} color=${args.color}
-      >The fox jumped over the lazy dog</wui-text
-    >`
+  tags: ['autodocs']
 } as Component
 
-export const Preview: Component = {
+export const Default: Component = {
   argTypes: {
     variant: {
       defaultValue: 'md-medium',
@@ -49,5 +44,9 @@ export const Preview: Component = {
       ],
       control: { type: 'select' }
     }
-  }
+  },
+  render: args =>
+    html`<wui-text variant=${args.variant} color=${args.color}
+      >The fox jumped over the lazy dog</wui-text
+    >`
 }
